Added more bindings
for the bindings in ed_io, I have modified the syntax with regards
to how the python functions are called. Python has no subroutines
and in general it's not able to modify arguments. If the arguments
are arrays though it is possible. That's why get_sigma and get_gimp
work. However, this is not optimal. I think we should change also
those.

For now, only the observable functions have been updated.

I have made it so that one can call e.g.

dens = ed.get_dens(#optional Nlat, #optional iorb)

in the driver.